Make assigning bugs only visible to project owner
